Update README.md
An AI based a robot navigation Agent using BFS, which can conduct blind searches to find its path from start to goal
state. As input, the system will take a description of the maize stored as a text file. The maize is a 2x2
grid with obstacles inside it. 
The obstacles are filled rectangles of unknown dimensions and can be found anywhere in the maize.
The robot cannot be in those cells. There are 3 actions allowed. Up one cell (cost is 1), right one cell
(cost is 3), diagonally up towards the right (cost is 2).